+++ index.php
@@ -28,6 +28,7 @@ function get_system_info(): array {
         'top'          => (string)(@shell_exec('ps aux | sort -rk 4,4 | head -n 11') ?: ''),
         'users'        => (string)(@shell_exec('who') ?: ''),
         'disks'        => (string)(@shell_exec('df -h') ?: ''),
+        'usb'          => (string)(@shell_exec('lsusb 2>/dev/null') ?: ''),
     ];
 }
 
@@ -516,6 +517,22 @@ function signal_quality(?int $dbm): string {
                 <div class="panel-heading"><i class="fa fa-hdd-o"></i> Disk Information</div>
                 <div class="panel-body">
                     <pre id="info-disks"><?php echo htmlspecialchars($sys['disks'], ENT_QUOTES, 'UTF-8'); ?></pre>
+                </div>
+            </div>
+        </div>
+    </div>
+
+    <!-- Connected USB devices -->
+    <div class="row">
+        <div class="col-md-12">
+            <div class="panel panel-default">
+                <div class="panel-heading"><i class="fa fa-usb"></i> Connected USB Devices</div>
+                <div class="panel-body">
+                    <?php if (trim($sys['usb']) === ''): ?>
+                        <pre id="info-usb" class="text-muted">No USB devices listed (<code>lsusb</code> not installed? <code>sudo apt-get install usbutils</code>)</pre>
+                    <?php else: ?>
+                        <pre id="info-usb"><?php echo htmlspecialchars($sys['usb'], ENT_QUOTES, 'UTF-8'); ?></pre>
+                    <?php endif; ?>
                 </div>
             </div>
         </div>